Place all analytics endpoints under the same file
